projects/capstone/starter/backend/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
distlib==0.3.4
ecdsa==0.18.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Moment==1.0.2
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-jose==3.3.0
pytz==2022.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.6
SQLAlchemy-Utils==0.38.2
typing_extensions==4.3.0
Werkzeug==2.0.1
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
